+++ tools/api-markdown-documenter/src/documentation-domain-to-markdown/configuration/Transformation.ts
@@ -6,8 +6,6 @@
 import type {
 	Nodes as MdastTree,
 	BlockContent as MdastBlockContent,
-	ListItem as MdastListItem,
-	PhrasingContent as MdastPhrasingContent,
 	RootContent as MdastRootContent,
 	TableCell as MdastTableCell,
 	TableRow as MdastTableRow,
@@ -15,30 +13,20 @@ import type {
 
 import type {
 	BlockContentMap,
-	PhrasingContentMap,
 	DocumentationNode,
 	SectionNode,
 	TableCellNode,
 	TableRowNode,
 	HeadingNode,
-	ListItemNode,
 } from "../../documentation-domain/index.js";
 import type { TransformationContext } from "../TransformationContext.js";
 import {
 	headingToMarkdown,
-	horizontalRuleToMarkdown,
 	sectionToMarkdown,
-	lineBreakToMarkdown,
-	paragraphToMarkdown,
-	plainTextToMarkdown,
-	spanToMarkdown,
 	tableToMarkdown,
 	tableCellToMarkdown,
 	tableRowToMarkdown,
-	listToMarkdown,
-	listItemToMarkdown,
 	markdownBlockContentNodeToMarkdown,
-	markdownPhrasingContentNodeToMarkdown,
 } from "../default-transformations/index.js";
 
 /**
@@ -54,30 +42,16 @@ export type BlockContentTransformations = {
 };
 
 /**
- * Transformations from {@link PhrasingContent} to {@link https://github.com/syntax-tree/mdast | Markdown syntax tree}s.
- *
- * @public
- */
-export type PhrasingContentTransformations = {
-	readonly [K in keyof PhrasingContentMap]: Transformation<
-		PhrasingContentMap[K],
-		MdastPhrasingContent[]
-	>;
-};
-
-/**
  * Transformations from {@link DocumentationNode}s to {@link https://github.com/syntax-tree/mdast | Markdown syntax tree}s.
  *
  * @public
  */
-export type Transformations = BlockContentTransformations &
-	PhrasingContentTransformations & {
-		readonly ["heading"]: Transformation<HeadingNode, MdastBlockContent[]>;
-		readonly ["listItem"]: Transformation<ListItemNode, [MdastListItem]>;
-		readonly ["section"]: Transformation<SectionNode, MdastRootContent[]>;
-		readonly ["tableCell"]: Transformation<TableCellNode, [MdastTableCell]>;
-		readonly ["tableRow"]: Transformation<TableRowNode, [MdastTableRow]>;
-	};
+export type Transformations = BlockContentTransformations & {
+	readonly ["heading"]: Transformation<HeadingNode, MdastBlockContent[]>;
+	readonly ["section"]: Transformation<SectionNode, MdastRootContent[]>;
+	readonly ["tableCell"]: Transformation<TableCellNode, [MdastTableCell]>;
+	readonly ["tableRow"]: Transformation<TableRowNode, [MdastTableRow]>;
+};
 
 /**
  * Transformation from a {@link DocumentationNode} to a {@link https://github.com/syntax-tree/mdast | Markdown syntax tree}.
@@ -97,16 +71,8 @@ export type Transformation<
  */
 export const defaultTransformations: Transformations = {
 	heading: headingToMarkdown,
-	lineBreak: lineBreakToMarkdown,
-	list: listToMarkdown,
-	listItem: listItemToMarkdown,
 	markdownBlockContent: markdownBlockContentNodeToMarkdown,
-	markdownPhrasingContent: markdownPhrasingContentNodeToMarkdown,
 	section: sectionToMarkdown,
-	horizontalRule: horizontalRuleToMarkdown,
-	paragraph: paragraphToMarkdown,
-	text: plainTextToMarkdown,
-	span: spanToMarkdown,
 	table: tableToMarkdown,
 	tableCell: tableCellToMarkdown,
 	tableRow: tableRowToMarkdown,