+++ tools/api-markdown-documenter/src/documentation-domain-to-markdown/default-transformations/index.ts
@@ -4,18 +4,8 @@
  */
 
 export { headingToMarkdown } from "./HeadingToMarkdown.js";
-export { horizontalRuleToMarkdown } from "./HorizontalRuleToMarkdown.js";
-export { lineBreakToMarkdown } from "./LineBreakToMarkdown.js";
-export { listToMarkdown } from "./ListToMarkdown.js";
-export { listItemToMarkdown } from "./ListItemToMarkdown.js";
-export {
-	markdownBlockContentNodeToMarkdown,
-	markdownPhrasingContentNodeToMarkdown,
-} from "./MarkdownNodeToMarkdown.js";
-export { paragraphToMarkdown } from "./ParagraphToMarkdown.js";
-export { plainTextToMarkdown } from "./PlainTextToMarkdown.js";
+export { markdownBlockContentNodeToMarkdown } from "./MarkdownNodeToMarkdown.js";
 export { sectionToMarkdown } from "./SectionToMarkdown.js";
-export { spanToMarkdown } from "./SpanToMarkdown.js";
 export { tableCellToMarkdown } from "./TableCellToMarkdown.js";
 export { tableRowToMarkdown } from "./TableRowToMarkdown.js";
 export { tableToMarkdown } from "./TableToMarkdown.js";