+++ tools/api-markdown-documenter/src/documentation-domain-to-markdown/test/DocumentToMarkdown.test.ts
@@ -9,10 +9,8 @@ import type { Root } from "mdast";
 import {
 	DocumentNode,
 	HeadingNode,
-	ParagraphNode,
-	PlainTextNode,
+	MarkdownBlockContentNode,
 	SectionNode,
-	SpanNode,
 } from "../../documentation-domain/index.js";
 import { documentToMarkdown } from "../ToMarkdown.js";
 
@@ -22,19 +20,43 @@ describe("documentToMarkdown", () => {
 			children: [
 				new SectionNode(
 					[
-						new ParagraphNode([
-							new PlainTextNode("This is a sample document. "),
-							new PlainTextNode("It has very basic content.\t"),
-						]),
+						new MarkdownBlockContentNode({
+							type: "paragraph",
+							children: [
+								{
+									type: "text",
+									value: "This is a sample document. ",
+								},
+								{
+									type: "text",
+									value: "It has very basic content.\t",
+								},
+							],
+						}),
 						new SectionNode(
 							[
-								new ParagraphNode([
-									new PlainTextNode("This is test inside of a paragraph. "),
-									new PlainTextNode("It is also inside of a hierarchical section node. "),
-									SpanNode.createFromPlainText("That's real neat-o.", {
-										italic: true,
-									}),
-								]),
+								new MarkdownBlockContentNode({
+									type: "paragraph",
+									children: [
+										{
+											type: "text",
+											value: "This is test inside of a paragraph. ",
+										},
+										{
+											type: "text",
+											value: "It is also inside of a hierarchical section node. ",
+										},
+										{
+											type: "emphasis",
+											children: [
+												{
+													type: "text",
+													value: "That's real neat-o.",
+												},
+											],
+										},
+									],
+								}),
 							],
 							new HeadingNode("Section Heading"),
 						),