+++ tools/api-markdown-documenter/src/documentation-domain/BlockContent.ts
@@ -3,10 +3,7 @@
  * Licensed under the MIT License.
  */
 
-import type { HorizontalRuleNode } from "./HorizontalRuleNode.js";
-import type { ListNode } from "./ListNode.js";
 import type { MarkdownBlockContentNode } from "./MarkdownNode.js";
-import type { ParagraphNode } from "./ParagraphNode.js";
 import type { TableNode } from "./TableNode.js";
 
 /**
@@ -33,9 +30,6 @@ import type { TableNode } from "./TableNode.js";
  * @public
  */
 export interface BlockContentMap {
-	horizontalRule: HorizontalRuleNode;
-	list: ListNode;
-	paragraph: ParagraphNode;
 	table: TableNode;
 	markdownBlockContent: MarkdownBlockContentNode;
 }