+++ tools/api-markdown-documenter/src/documentation-domain/MarkdownNode.ts
@@ -3,10 +3,7 @@
  * Licensed under the MIT License.
  */
 
-import type {
-	BlockContent as MdastBlockContent,
-	PhrasingContent as MdastPhrasingContent,
-} from "mdast";
+import type { BlockContent as MdastBlockContent } from "mdast";
 
 import { DocumentationLiteralNodeBase } from "./DocumentationNode.js";
 
@@ -27,43 +24,7 @@ export class MarkdownBlockContentNode extends DocumentationLiteralNodeBase<Mdast
 	 */
 	public readonly type = "markdownBlockContent";
 
-	/**
-	 * {@inheritDoc DocumentationNode.isEmpty}
-	 */
-	public get isEmpty(): boolean {
-		return false; // Not well defined
-	}
-
 	public constructor(value: MdastBlockContent) {
-		super(value);
-	}
-}
-
-/**
- * A {@link DocumentationNode} that wraps an `mdast` "Phrasing content" tree.
- *
- * @remarks
- *
- * This type exists as a temporary solution while this library is migrated to using `mdast` directly (replacing the entire `DocumentationNode` domain).
- * It will eventually be removed, along with the rest of this domain.
- *
- * @sealed
- * @public
- */
-export class MarkdownPhrasingContentNode extends DocumentationLiteralNodeBase<MdastPhrasingContent> {
-	/**
-	 * {@inheritDoc DocumentationNode."type"}
-	 */
-	public readonly type = "markdownPhrasingContent";
-
-	/**
-	 * {@inheritDoc DocumentationNode.isEmpty}
-	 */
-	public get isEmpty(): boolean {
-		return false; // Not well defined
-	}
-
-	public constructor(value: MdastPhrasingContent) {
 		super(value);
 	}
 }