+++ tools/api-markdown-documenter/src/documentation-domain/index.ts
@@ -27,16 +27,8 @@ export {
 	DocumentationParentNodeBase,
 } from "./DocumentationNode.js";
 export { HeadingNode } from "./HeadingNode.js";
-export { HorizontalRuleNode } from "./HorizontalRuleNode.js";
-export { LineBreakNode } from "./LineBreakNode.js";
-export { ListItemNode } from "./ListItemNode.js";
-export { ListNode } from "./ListNode.js";
-export { MarkdownBlockContentNode, MarkdownPhrasingContentNode } from "./MarkdownNode.js";
-export { ParagraphNode } from "./ParagraphNode.js";
-export type { PhrasingContent, PhrasingContentMap } from "./PhrasingContent.js";
-export { PlainTextNode } from "./PlainTextNode.js";
+export { MarkdownBlockContentNode } from "./MarkdownNode.js";
 export { type SectionContent, SectionNode } from "./SectionNode.js";
-export { SpanNode } from "./SpanNode.js";
 export {
 	type TableCellContent,
 	TableCellNode,
@@ -51,5 +43,3 @@ export {
 	TableRowKind,
 } from "./TableRowNode.js";
 export { TableNode } from "./TableNode.js";
-export type { TextFormatting } from "./TextFormatting.js";
-export { createNodesFromPlainText } from "./Utilities.js";