+++ examples/apps/presence-tracker/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @fluid-example/presence-tracker
 
+## 2.52.0
+
+Dependency updates only.
+
 ## 2.51.0
 
 Dependency updates only.